requirements.txt
openai-whisper>=2023.8.8
torch
tqdm
flask
werkzeug
